requirements.txt
yfinance==0.2.36
pandas==2.1.0
scikit-learn==1.3.0
fastapi==0.104.0
uvicorn==0.23.2
python-dotenv==1.0.0
Flask==2.0.1
fpdf2==2.7.6
textblob==0.17.1